move tagset to a new pkg
This moves the following tagset-related functionality into a dedicated
package:

 - TagsBuilder
 - HashingTagsBuidler
 - HashedTags

It adds a few tests, but represents no change in functionality.
