+++ pkg/aggregator/ckey/key.go
@@ -8,7 +8,7 @@ package ckey
 import (
 	"math/bits"
 
-	"github.com/DataDog/datadog-agent/pkg/util"
+	"github.com/DataDog/datadog-agent/pkg/tagset"
 	"github.com/twmb/murmur3"
 )
 
@@ -75,7 +75,7 @@ type KeyGenerator struct {
 
 // Generate returns the ContextKey hash for the given parameters.
 // tagsBuf is re-arranged in place and truncated to only contain unique tags.
-func (g *KeyGenerator) Generate(name, hostname string, tagsBuf *util.HashingTagsBuilder) ContextKey {
+func (g *KeyGenerator) Generate(name, hostname string, tagsBuf *tagset.HashingTagsBuilder) ContextKey {
 	// between two generations, we have to set the hash to something neutral, let's
 	// use this big value seed from the murmur3 implementations
 	g.intb = 0xc6a4a7935bd1e995
