+++ pkg/aggregator/time_sampler_test.go
@@ -18,12 +18,12 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/aggregator/ckey"
 	"github.com/DataDog/datadog-agent/pkg/metrics"
 	"github.com/DataDog/datadog-agent/pkg/quantile"
-	"github.com/DataDog/datadog-agent/pkg/util"
+	"github.com/DataDog/datadog-agent/pkg/tagset"
 )
 
 func generateSerieContextKey(serie *metrics.Serie) ckey.ContextKey {
 	l := ckey.NewKeyGenerator()
-	return l.Generate(serie.Name, serie.Host, util.NewHashingTagsBuilderWithTags(serie.Tags))
+	return l.Generate(serie.Name, serie.Host, tagset.NewHashingTagsBuilderWithTags(serie.Tags))
 }
 
 // TimeSampler
@@ -320,7 +320,7 @@ func TestSketch(t *testing.T) {
 					Ts:     0,
 				},
 			},
-			ContextKey: keyGen.Generate(ctx.Name, ctx.Host, util.NewHashingTagsBuilderWithTags(ctx.Tags)),
+			ContextKey: keyGen.Generate(ctx.Name, ctx.Host, tagset.NewHashingTagsBuilderWithTags(ctx.Tags)),
 		}, flushed[0])
 
 		_, flushed = sampler.flush(now)
