+++ pkg/metrics/histogram_bucket.go
@@ -5,7 +5,7 @@
 
 package metrics
 
-import "github.com/DataDog/datadog-agent/pkg/util"
+import "github.com/DataDog/datadog-agent/pkg/tagset"
 
 // HistogramBucket represents a prometheus/openmetrics histogram bucket
 type HistogramBucket struct {
@@ -33,7 +33,7 @@ func (m *HistogramBucket) GetHost() string {
 }
 
 // GetTags returns the bucket tags.
-func (m *HistogramBucket) GetTags(tb *util.HashingTagsBuilder) {
+func (m *HistogramBucket) GetTags(tb *tagset.HashingTagsBuilder) {
 	// Other 'GetTags' methods for metrics support origin detections. Since
 	// HistogramBucket only come, for now, from checks we can simply return
 	// tags.
