+++ pkg/metrics/test_helper.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/pkg/aggregator/ckey"
 	"github.com/DataDog/datadog-agent/pkg/quantile"
-	"github.com/DataDog/datadog-agent/pkg/util"
+	"github.com/DataDog/datadog-agent/pkg/tagset"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -163,7 +163,7 @@ func Makeseries(i int) SketchSeries {
 	}
 
 	gen := ckey.NewKeyGenerator()
-	ss.ContextKey = gen.Generate(ss.Name, ss.Host, util.NewHashingTagsBuilderWithTags(ss.Tags))
+	ss.ContextKey = gen.Generate(ss.Name, ss.Host, tagset.NewHashingTagsBuilderWithTags(ss.Tags))
 
 	return ss
 }
